🚧 Change how effects are defined
